seperate deflection calculations into testable units
